requirements/base.txt
# base.txt is a pip requirements file which describes the necessary
# distributions required to run this project. If you need something that is
# project specific not listed here use project.txt. You can, of course, update
# versions of distributions here if needed.

--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/
--extra-index-url=http://dist.pinaxproject.com/fresh-start/

BeautifulSoup==3.2.1
Django==1.4.1
Jinja2==2.6

Pinax==0.9b1.dev10
Pygments==1.5
South==0.7.4
Sphinx==1.1.3
certifi==0.0.8
chardet==1.0.1
cmsplugin-filer==0.8.0
decorator==3.3.2
django-admin-tools==0.4.1
django-announcements==0.1.2
django-appconf==0.5
django-classy-tags==0.3.4.1
django-cms==2.2
django-compressor==1.1.2
django-debug-toolbar==0.8.5
django-email-confirmation==0.2
django-extensions==0.8
git+https://github.com/stefanfoulis/django-filer.git@f71772a274c1d2d4b59e5b3c999b19f2826f5253#egg=django-filer
git+https://github.com/toastdriven/django-haystack.git@master#egg=django-haystack
django-mailer==0.2a1
django-modeltranslation==0.3.3
django-mptt==0.5.2
django-notification==0.2
django-openid==0.3a1
django-pagination==1.0.7
django-polymorphic==0.2
django-redis-cache==0.9.2
django-sekizai==0.5
git+https://github.com/festlv/django-social-auth.git
git+https://github.com/festlv/sorl-watermark.git#egg=sorl-watermark
git+https://github.com/festlv/gunicorn.git
git+https://github.com/festlv/django-activity-stream.git
django-staticfiles==1.2.1
django-timezones==0.2
docutils==0.8.1
easy-thumbnails==1.0-alpha-21


html5lib==0.95
httplib2==0.7.4
idios==0.2a1
ipcalc==0.5

metron==0.1
oauth2==1.5.211
pinax-theme-bootstrap==0.1.2
python-openid==2.2.5
pytz==2011n
redis==2.4.11
requests==0.11.1
sorl-thumbnail==11.12
urllib3==1.3
versiontools==1.9.1
wsgiref==0.1.2
selenium
django-selenium
